==> 4 - Compute correlations.ipynb <==
import pandas as pd

path = './new_cleaned'

topic_df_statistics = pd.read_csv(f'{path}/topics_all_statistics_and_scores.csv')

topic_df_statistics = topic_df_statistics.drop(topic_df_statistics[topic_df_statistics.last_3_month_completed_step_users_count< 20].index)

topic_df_statistics.head()

topic_df_statistics.shape

topic_df_statistics = topic_df_statistics.dropna()

topic_df_statistics.shape  # there were no Nan values

# compute average number of sentences and words in topics
only_two = topic_df_statistics[['num_sentences', 'num_words']]

overall = only_two.shape[0]
sents = 0
words = 0

for i, row in only_two.iterrows():
    sents += row['num_sentences']
    words += row['num_words']
    
print('sents', sents, round(sents/overall, 2))
print('words', words, round(words/overall, 2))

for i, row in only_two.iterrows():
    print(row['num_sentences'])
    break

only_two.shape

## Compute correlations

Correlations between Flesch score / Dale-Chall score and:

* normalized seconds_to_complete
* last_3_month_completion_rate
* last_3_month_avg_like (but only for those where last_year_likes_count is > 20)
* last_3_month_topic_completion_rate
* back_to_theory_times_per_user_session_avg_last_3_month
* back_to_theory_users_%_last_3_month

from scipy.stats import spearmanr

def correlate_two(df, name_row1, name_row2):    
    score, pvalue = spearmanr(a=df[name_row1].values, b=df[name_row2].values, nan_policy='raise')
    return score, pvalue

print('FLESCH')
for column in ['norm_seconds', 'last_3_month_completion_rate', 'last_3_month_topic_completion_rate', 
              'back_to_theory_times_per_user_session_avg_last_3_month', 'back_to_theory_users_%_last_3_month']:
    print(column)
    
    score, p = correlate_two(topic_df_statistics, 'flesch_score', column)
    print(round(score, 3), round(p, 5))
    print()

# for last_year_avg_like we need to take rows where last_year_likes_count > 20
df_with_likes = topic_df_statistics.drop(topic_df_statistics[topic_df_statistics.last_3_month_likes_count < 20].index)

score, p = correlate_two(df_with_likes, 'flesch_score', 'last_3_month_avg_like')
print(round(score, 3), round(p, 5))

print('DALE-CHALL')
for column in ['norm_seconds', 'last_3_month_completion_rate', 'last_3_month_topic_completion_rate', 
              'back_to_theory_times_per_user_session_avg_last_3_month', 'back_to_theory_users_%_last_3_month']:
    print(column)
    
    score, p = correlate_two(topic_df_statistics, 'dale_chall_score', column)
    print(round(score, 3), round(p, 5))
    print()

score, p = correlate_two(df_with_likes, 'dale_chall_score', 'last_3_month_avg_like')
print(round(score, 3), round(p, 5))

# the scores between each other
score, p = correlate_two(topic_df_statistics, 'dale_chall_score', 'flesch_score')
print(round(score, 3), round(p, 5))

# and now -- between all the parameters with each other
for column1 in ['norm_seconds', 'last_3_month_completion_rate', 'last_3_month_topic_completion_rate', 
               'last_3_month_avg_like', 'back_to_theory_times_per_user_session_avg_last_3_month', 
               'back_to_theory_users_%_last_3_month']:
    for column2 in ['norm_seconds', 'last_3_month_completion_rate', 'last_3_month_topic_completion_rate', 
                    'last_3_month_avg_like', 'back_to_theory_times_per_user_session_avg_last_3_month',
                    'back_to_theory_users_%_last_3_month']:
        if column1 == column2:
            continue
        score, p = correlate_two(df_with_likes, column1, column2)
        print(column1, column2)
        print(round(score, 3), round(p, 5))

## Correlations with other readability scores

path = '/home/anna/Documents/Diploma/code/new_cleaned'

df_with_all_scores = pd.read_csv(f'{path}/topics_20_compl_statistics_and_MORE_scores.csv')

for score in ['flesch_kincaid', 'gunning_fog', 'coleman_liau', 'ari', 'linsear_write', 'spache']:
    print('###', score, '###')
    for column in ['norm_seconds', 'last_3_month_completion_rate', 'last_3_month_topic_completion_rate', 
                   'last_3_month_avg_like', 'back_to_theory_times_per_user_session_avg_last_3_month', 
                   'back_to_theory_users_%_last_3_month']:
        print(column)

        index, p = correlate_two(df_with_all_scores, f'{score}_score', column)
        print(round(index, 3), round(p, 5))
        print()
    print('##########')

# for smog we need to disregard those values where there are zeros
df_topics_for_smog = df_with_all_scores.drop(df_with_all_scores[df_with_all_scores.smog_score == 0].index)

for score in ['smog']:
    print(score)
    for column in ['norm_seconds', 'last_3_month_completion_rate', 'last_3_month_topic_completion_rate', 
                   'last_3_month_avg_like', 'back_to_theory_times_per_user_session_avg_last_3_month', 
                   'back_to_theory_users_%_last_3_month']:
        print(column)

        index, p = correlate_two(df_topics_for_smog, f'{score}_score', column)
        print(round(index, 3), round(p, 5))
        print()
    print('##########')
