==> 5 - Get lists of topics' prerequisites.ipynb <==
## Collect all prerequisites for theory steps

import pandas as pd

import requests

topic_statistics = '../topics_with_statistics/bigger_statistics_one_year.csv'

one_year_statistics_df = pd.read_csv(topic_statistics)

one_year_statistics_df.head()

topic_statistics = one_year_statistics_df.drop(one_year_statistics_df[one_year_statistics_df.is_theory == 0].index)

topic_statistics.step_id.describe()

def get_topic_id(step_id, link='https://hyperskill.org/api'):
    response = requests.get(f'{link}/steps/{step_id}')
    return response.json()['steps'][0]['topic']

def get_immediate_prereq(topic_id, link='https://hyperskill.org/api'):
    response = requests.get(f'{link}/topics/{topic_id}')
    try:
        result = response.json()['topics'][0]['prerequisites']
    except:
        print(response.json())
        raise
    return result

def get_all_prereq(topic_id, dict_with_topics_info, link='https://hyperskill.org/api', _print=False):
    if topic_id in dict_with_topics_info:
        if _print: print(f'topic {topic_id} is already in dict')
        return dict_with_topics_info[topic_id]
    try:
        list_prereq = get_immediate_prereq(topic_id)
        if _print: print(f'immediate prereq of {topic_id} are {list_prereq}')
    except:
        print(topic_id)
        if topic_id in {1032, 350, 1075, 1174, 1114}:  # these topics are not found
            if _print: print(f'passing because of {topic_id}')
            list_prereq = []
            pass
        else:
            raise
    if not list_prereq:  # list is empty
        dict_with_topics_info[topic_id] = []
        if _print: print(f'prereq for {topic_id} are empty')
        return []
    
    set_prereq = set(list_prereq)
        
    for el in list_prereq:
        if _print: print(f'getting prereq of {el}')
            
        new_prereq = get_all_prereq(el, dict_with_topics_info, _print=_print)  # recursion
        dict_with_topics_info[el] = new_prereq
        
        set_prereq = set_prereq.union(new_prereq)
    return list(set_prereq)      

step_id_topic_id_mapping = {}
topic_id_step_id_mapping = {}

for i, s_id in enumerate(topic_statistics.step_id.values):
    if i % 10 == 0:
        print(i)
        
    corresponding_topic_id = get_topic_id(s_id)
    
    step_id_topic_id_mapping[s_id] = corresponding_topic_id
    topic_id_step_id_mapping[corresponding_topic_id] = s_id

import pickle

pickle.dump(step_id_topic_id_mapping, open("step_id_topic_id_mapping.pkl", "wb"))
pickle.dump(topic_id_step_id_mapping, open("topic_id_step_id_mapping.pkl", "wb"))

### Collect all prerquisites

#dict_w_topic_prerequisites = dict()

for i, topic_id in enumerate(topic_id_step_id_mapping):
    if topic_id in dict_w_topic_prerequisites:
        continue
    print(i)
        
    prerequisites = get_all_prereq(topic_id, dict_w_topic_prerequisites)
    dict_w_topic_prerequisites[topic_id] = prerequisites

# 1032, 350, 1075, 1174, 1114 -- errors

get_all_prereq(1032, dict(), _print=True)

pickle.dump(dict_w_topic_prerequisites, open("topic_id_all_prereq_mapping.pkl", "wb" ))

### Collect direct prerequisites

direct_prereqs = dict()

for i, s_id in enumerate(topic_statistics.step_id.values):
        
    if i % 10 == 0:
        print(i)
        
    corresponding_topic_id = step_id_topic_id_mapping[s_id]
    
    if corresponding_topic_id in direct_prereqs:
        continue
    try:
        direct_prerequisites = get_immediate_prereq(corresponding_topic_id)
    except:
        print(s_id, corresponding_topic_id)
    
    direct_prereqs[corresponding_topic_id] = direct_prerequisites
    

# for these there was an error with 'not found'
print(direct_prereqs[1032])  # most probably there's the prerequisites of a previous topic
print(direct_prereqs[350])   # need to delete them from the dictionary
print(direct_prereqs[1075])
print(direct_prereqs[1174])  
print(direct_prereqs[1114])

# for these the problem was probably temporary
print(direct_prereqs[1542])
print(direct_prereqs[1081])
print(direct_prereqs[1468])
print(direct_prereqs[1625])

for topic_id in [1032, 350, 1075, 1174, 1114]:  # for these there was an error with 'not found'
    print(topic_id)
    try:
        direct_prerequisites = get_immediate_prereq(topic_id)
        print(direct_prerequisites)
    except:
        print('error')
        
# these are deleted topics that were mentioned as prerequisites, they are now disabled

for topic_id in [1542, 1081, 1468, 1625]:  # for these the problem was probably temporary
    print(topic_id)
    try:
        direct_prerequisites = get_immediate_prereq(topic_id)
        print(direct_prerequisites)
    except:
        print('error')

# now we delete from the dictionary those that had 'not found' errors
# and update values for the two with the other kind of error
for topic in [1032, 350, 1075, 1174, 1114]:
    del direct_prereqs[topic]
    
for topic in [1542, 1081, 1468, 1625]:
    direct_prerequisites = get_immediate_prereq(topic)
    direct_prereqs[topic] = direct_prerequisites

pickle.dump(direct_prereqs, open("topic_id_direct_prereq_mapping.pkl", "wb" ))
