==> 10 - Collect features for ML.ipynb <==
We will prepare features for all topics, disregarding the 'target' features. Later from the DFs with features we'll take only those topics that build our training corpus.

## Take features from Quita Up

import pandas as pd

quita_features_file = './new_cleaned/all_quita_features_all_texts.csv'

quita_features = pd.read_csv(quita_features_file)
quita_features.head()

# from this we don't need 'VarEntropy_lemma', 'descriptivity' 
# 'Text' is step id

#### See correlations between the features

from scipy import stats

# see if the 3 mattr attributes correlate
coef1, pval1 = stats.spearmanr(quita_features['mattr_lemma_100'], quita_features['mattr_lemma_50'])
coef2, pval2 = stats.spearmanr(quita_features['mattr_lemma_100'], quita_features['mattr_lemma_25'])
coef3, pval3 = stats.spearmanr(quita_features['mattr_lemma_50'], quita_features['mattr_lemma_25'])

print(coef1, pval1)
print(coef2, pval2)
print(coef3, pval3)

# they do, so we'll take only one of them (50)

list_taking = ['hpoint_lemma', 'Entropy_lemma', 'verbdistance_lemma', 'activity', 'ATL_token', 'TC_lemma', 
               'mattr_lemma_50', 'zttr_FIClemma']

for i, attribute1 in enumerate(list_taking):
    for attribute2 in list_taking[i+1:]:
        coef, pval = stats.spearmanr(quita_features[attribute1], quita_features[attribute2])
        
        if abs(coef) > 0.5 and pval < 0.05: 
            print(attribute1, attribute2)
            print(round(coef, 2), round(pval, 2))
            print()

zttr_FIClemma correlates a lot with Entropy_lemma and mattr, so we'll not take it among the attributes

#### Create a DF with quita features for topic IDs

only_needed_features = ['topic_id', 'hpoint_lemma', 'Entropy_lemma', 'verbdistance_lemma', 'activity', 
                        'ATL_token', 'TC_lemma', 'mattr_lemma_50']

quita_needed_df = quita_features[only_needed_features]
quita_needed_df.head()

quita_needed_df.to_csv('./new_cleaned/train_corpus/quita_features.csv', index=False)

### Collect other linguistic features

# DC score
# Flesch score
# average sentence length

import pandas as pd

import pickle

topic_id_step_id_mapping = pickle.load(open('topic_id_step_id_mapping.pkl', 'br'))

step_id_topic_id_mapping = pickle.load(open('step_id_topic_id_mapping.pkl', 'br'))

topics_statistics = pd.read_csv('./new_cleaned/topics_all_statistics_and_scores.csv')

topics_statistics.head()

ling_features_df = quita_needed_df.copy()

# in ling features there's now one row more than in topic statistics; we need to delete this row

print(ling_features_df.shape[0])

bad_step_id = 16324
bad_topic_id = step_id_topic_id_mapping[bad_step_id]
ling_features_df = ling_features_df.drop(ling_features_df[ling_features_df.topic_id == bad_topic_id].index)
print(ling_features_df.shape[0])

for attr in ['dale_chall_score', 'flesch_score', 'ASL']:
    attr_list = []

    for row in ling_features_df.iterrows():
        topic_id = row[1].topic_id
        step_id = topic_id_step_id_mapping[topic_id]

        corresponding_row = topics_statistics[topics_statistics.step_id == step_id]
        assert not corresponding_row.empty, f'In topic statistics there\'s no row with step_id {step_id}'

        attr_value = corresponding_row[attr].values[0]
        attr_list.append(attr_value)

    ling_features_df[attr] = attr_list

ling_features_df.head()

ling_features_df.to_csv('./new_cleaned/train_corpus/all_ling_features.csv', index=False)

### Add probability of the text from GPT-2

#### Try with the first example

#!pip3 install lm-scorer

import torch
from lm_scorer.models.auto import AutoLMScorer as LMScorer

import pandas as pd

list(LMScorer.supported_model_names())

# Load model to cpu or cuda
available_gpu = torch.cuda.is_available()
print(available_gpu)
device = "cuda:0" if available_gpu else "cpu"

#!pip3 install ipywidgets

batch_size = 1
scorer = LMScorer.from_pretrained("gpt2", device=device, batch_size=batch_size)

# Return token probabilities
scorer.tokens_score("I like this package.")  # the higher, the more probable

scorer.tokens_score("I like this package.", log=True)  # the higher, the more probable

# we don't want to use multiplication of tokens' probabilities
# cause 1) results will be very small numbers, 2) it will be influences by the number of tokens

# Compute sentence score as the geometric mean of tokens' probabilities
print(scorer.sentence_score("I like this package.", reduce="gmean"))

# Get the log of the sentence score
print(scorer.sentence_score("I like this package.", log=True))

# Score multiple sentences
print(scorer.sentence_score(["Sentence 1", "Sentence 2"]))
print(scorer.sentence_score(["What a wonderful day", "What a wonderful life"], reduce="gmean"))

print(scorer.sentence_score(["Sentence 1", "Sentence 2"], reduce='gmean', log=True))

#### Get scores for sentences in each topic

import pickle

import spacy
en_sm_model = spacy.load("en_core_web_sm")

raw_text = 'Hello, world. Here are two sentences.'
doc = en_sm_model(raw_text)

sentences = [sent.text.strip() for sent in doc.sents]
sentences

def get_spacy_sents(text, model):
    doc = model(text)
    sentences = [sent.text.strip() for sent in doc.sents]
    return sentences

# since we score sentences, we need to get sentences for each topic
# for now we have cleaned text and lemmatized text
# we'll need to go through cleaned texts again, split them into sentences and receive sentence scores

topics_statistics = pd.read_csv('./new_cleaned/topics_all_statistics_and_scores.csv')

topics_statistics.head()

topics_statistics.shape

def delete_special_tokens(text):
    for tok in ['*heading*', '*img*', '*code*', '*math*', '*table*']:
        text = text.replace(tok, '')
    return text

# for each topic (step id) we get a list of prbabilities of its sentences
step_id_sent_prob_mapping = dict()

for i, row in topics_statistics.iterrows():
    print(i)
    cleaned = row.cleaned_texts
    cleaned = delete_special_tokens(cleaned)
    
    # split into sentences
    sentences = get_spacy_sents(cleaned, en_sm_model)
    
    # compute probability of sentences
    sent_probs = scorer.sentence_score(sentences, reduce="gmean")
    
    step_id_sent_prob_mapping[row.step_id] = sent_probs

step_id_sent_prob_mapping[12691]

pickle.dump(step_id_sent_prob_mapping, open('./new_cleaned/train_corpus/step_id_sent_prob_mapping.pkl', 'wb'))

### From sentence scores, get a score for each topic

import statistics

# now we prepare to write them into a DF

df_with_topic_id = pd.read_csv('./new_cleaned/train_corpus/all_ling_features.csv')
topic_ids = df_with_topic_id.topic_id.values
print(len(topic_ids))

topic_id_step_id_map = pickle.load(open('./topic_id_step_id_mapping.pkl', 'rb'))

step_id_sent_prob_mapping = pickle.load(open('./new_cleaned/train_corpus/step_id_sent_prob_mapping.pkl', 'rb'))

def get_topic_score_from_sentence_scores(topic_id, topic_id_step_id_mapping, step_id_sent_prob_mapping):
    corr_step_id = topic_id_step_id_mapping[topic_id]
    list_sent_scores = step_id_sent_prob_mapping[corr_step_id]
    return statistics.mean(list_sent_scores)

LM_scores = []
for topic_id in topic_ids:
    LM_scores.append(get_topic_score_from_sentence_scores(topic_id, topic_id_step_id_mapping, step_id_sent_prob_mapping))

lm_prob_df = pd.DataFrame({'topic_id': topic_ids, 'GPT2_prob': LM_scores})

lm_prob_df.head()

lm_prob_df.to_csv('./new_cleaned/train_corpus/LM_feature.csv', index=False)

## Collect meta-features

### Preparations

# the number of overall prereqs - from topic_id_all_prereq_mapping.pkl
# the number of direct prereqs - from topic_id_direct_prereq_mapping.pkl
# the number of code snippets - the number of *code*

# the number of images - need to replace images to *img* in text extraction
# the number of sections - need to count headings in text extraction

# percent of code in the whole text (in symbols)
# the topic group -- from the comments' dump -- DO NOT TAKE

import pickle

import pandas as pd

topic_id_all_prereq_mapping = pickle.load(open('topic_id_all_prereq_mapping.pkl', 'rb'))
topic_id_direct_prereq_mapping = pickle.load(open('topic_id_direct_prereq_mapping.pkl', 'rb'))
step_id_topic_id_mapping = pickle.load(open('step_id_topic_id_mapping.pkl', 'br'))
topic_id_step_id_mapping = pickle.load(open('topic_id_step_id_mapping.pkl', 'br'))

topics_statistics = pd.read_csv('./new_cleaned/topics_all_statistics_and_scores.csv')

topics_statistics.columns

def get_overall_prereqs(topic_id, topic_id_all_prereq_mapping):
    return len(topic_id_all_prereq_mapping[topic_id])

def get_direct_prereqs(topic_id, topic_id_direct_prereq_mapping):
    return len(topic_id_direct_prereq_mapping[topic_id])

def get_num_code_snippets(topic_id, topic_statistics_df, topic_id_step_id_mapping):
    step_id = topic_id_step_id_mapping[topic_id]
    row = topic_statistics_df[topic_statistics_df.step_id == step_id]
    text = row.cleaned_texts.values[0]
    return text.count("*code*")

def get_num_images(topic_id, topic_statistics_df, topic_id_step_id_mapping):
    step_id = topic_id_step_id_mapping[topic_id]
    row = topic_statistics_df[topic_statistics_df.step_id == step_id]
    text = row.cleaned_texts.values[0]
    return text.count("*img*")

def get_num_sections(topic_id, topic_statistics_df, topic_id_step_id_mapping):
    step_id = topic_id_step_id_mapping[topic_id]
    row = topic_statistics_df[topic_statistics_df.step_id == step_id]
    return row.num_headings.values[0]

import re
pattern = re.compile(r'\s+')

def get_percent_code(topic_id, topic_statistics_df, topic_id_step_id_mapping):  
    # percent of code in symbols, wo spaces
    step_id = topic_id_step_id_mapping[topic_id]
    row = topic_statistics_df[topic_statistics_df.step_id == step_id]
    text = row.cleaned_texts.values[0]
    wo_spaces = re.sub(pattern, '', text)
    
    snippets_len = row.symbols_in_snippets.values[0]
    return snippets_len / (len(wo_spaces) + snippets_len)

def collect_all_features(topic_id, topic_statistics_df, topic_id_step_id_mapping, 
                        topic_id_all_prereq_mapping, topic_id_direct_prereq_mapping):
    return (get_overall_prereqs(topic_id, topic_id_all_prereq_mapping),
            get_direct_prereqs(topic_id, topic_id_direct_prereq_mapping),
            get_num_code_snippets(topic_id, topic_statistics_df, topic_id_step_id_mapping),
            get_num_images(topic_id, topic_statistics_df, topic_id_step_id_mapping),
            get_num_sections(topic_id, topic_statistics_df, topic_id_step_id_mapping),
            get_percent_code(topic_id, topic_statistics_df, topic_id_step_id_mapping))

### Create a df with meta features

df_with_topic_id = pd.read_csv('./new_cleaned/train_corpus/all_ling_features.csv')
topic_ids = df_with_topic_id.topic_id.values
print(len(topic_ids))

print(len(topic_id_direct_prereq_mapping), len(topic_id_all_prereq_mapping))

# check what topics are not in prereq mappings
ids_not_in_mapping = []

for t_id in topic_ids:
    if t_id not in topic_id_direct_prereq_mapping:
        print(t_id)
        ids_not_in_mapping.append(t_id)
        
print('now all prereqs')

for t_id in topic_ids:
    if t_id not in topic_id_all_prereq_mapping:
        print(t_id)

# for the alg to work without mistakes, let's delete these topic IDs from topic_ids
topic_ids = list(topic_ids)
print(len(topic_ids))

for el in ids_not_in_mapping:
    topic_ids.remove(el)
    
print(len(topic_ids))

dict_features = dict()
features_list = ['overall_prereqs', 'direct_prereqs', 'num_snippets', 'num_img', 'num_sections', 'percent_code']

for feature in features_list:
    dict_features[feature] = []

for topic_id in topic_ids:
        six_values = collect_all_features(topic_id, topics_statistics, topic_id_step_id_mapping,
                                          topic_id_all_prereq_mapping, topic_id_direct_prereq_mapping)
        
        for key, value in zip(dict_features.keys(), six_values):
            dict_features[key].append(value)

df_meta_features = pd.DataFrame(columns=['topic_id'] + features_list)

df_meta_features['topic_id'] = topic_ids
for feature in features_list:
    df_meta_features[feature] = dict_features[feature]

df_meta_features.head()

df_meta_features.to_csv('./new_cleaned/train_corpus/meta_features.csv', index=False)

## Add statistical features

# we'll need to disregard average like if it was used to extend the corpus

import pandas as pd

import pickle

df_with_topic_id = pd.read_csv('./new_cleaned/train_corpus/all_ling_features.csv')
topic_ids = df_with_topic_id.topic_id.values
print(len(topic_ids))

topics_statistics = pd.read_csv('./new_cleaned/topics_all_statistics_and_scores.csv')

print(len(topic_ids), topics_statistics.shape[0])

step_id_topic_id_mapping = pickle.load(open('step_id_topic_id_mapping.pkl', 'br'))

topics_statistics.columns

# we take only one of the two back_to_theory cause they correlate a lot (0.97)
needed_features_w_like = ['norm_seconds', 'last_3_month_completion_rate', 'last_3_month_avg_like',
                          'last_3_month_topic_completion_rate', 'back_to_theory_users_%_last_3_month']
needed_features_wo_like = ['norm_seconds', 'last_3_month_completion_rate',
                          'last_3_month_topic_completion_rate', 'back_to_theory_users_%_last_3_month']

topics_statistics_with_completions = topics_statistics[topics_statistics.last_3_month_completed_step_users_count > 20]
topics_statistics_with_likes = topics_statistics[topics_statistics.last_3_month_likes_count > 20]

def fill_in_df(topic_statistics, needed_features, step_id_topic_id_mapping):
    dict_w_features = dict()
    for feature in ['topic_id'] + needed_features:
        dict_w_features[feature] = []
    
    for i, row in topic_statistics.iterrows():
        corr_topic_id = step_id_topic_id_mapping[row.step_id]

        dict_w_features['topic_id'].append(corr_topic_id)
        for feature in needed_features:
            dict_w_features[feature].append(row[feature])  # values[0]

    df_result = pd.DataFrame(dict_w_features)
    return df_result

df_with_likes = fill_in_df(topics_statistics_with_likes, needed_features_w_like, step_id_topic_id_mapping)
assert df_with_likes.shape[0] == topics_statistics_with_likes.shape[0]
df_with_likes.head()

df_wo_likes = fill_in_df(topics_statistics_with_completions, needed_features_wo_like, step_id_topic_id_mapping)
assert df_wo_likes.shape[0] == topics_statistics_with_completions.shape[0]
df_wo_likes.head()

df_with_likes.to_csv('./new_cleaned/train_corpus/statistics_with_likes.csv', index=False)
df_wo_likes.to_csv('./new_cleaned/train_corpus/statistics_without_likes.csv', index=False)
